% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delury.R
\name{delury}
\alias{delury}
\title{Delury removal model}
\usage{
delury(dat)
}
\arguments{
\item{dat}{dat_remove or a similarly formatted list.}
}
\value{
An object of class lm
}
\description{
Runs the Delury estimate removal estimate
}
\examples{
delury(dat_remove[[1]])
lapply(dat_remove, delury)

}
