% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie.R
\name{leslie}
\alias{leslie}
\title{Leslie removal model}
\usage{
leslie(dat)
}
\arguments{
\item{dat}{dat_remove or a similarly formatted list.}
}
\value{
An object of class lm
}
\description{
Runs the Leslie estimate removal estimate
}
\examples{
leslie(dat_remove[[1]])
lapply(dat_remove, leslie)

}
