% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_df.R
\name{remove_df}
\alias{remove_df}
\title{remove_df}
\usage{
remove_df(data, i, n = names(data))
}
\arguments{
\item{data}{dat_remove or a similarly formatted list.}

\item{i}{elements of dat_remove to include, i.e. dat_remove[[i]]}

\item{n}{experiment names, defaults to names(dat_remove[i])}
}
\description{
Creates a data frame from the dat_remove list. Can subset dat_remove with i.
}
\examples{
remove_df(i = 1, dat_remove)
plyr::ldply(1:12, remove_df, data = dat_remove)
plyr::ldply(c(3, 7), remove_df, data = dat_remove)

}
